00 days of code

+++ Q100.c
@@ -4,10 +4,10 @@
 int main() {
     char str[1000];
     scanf("%s", str);
-    int len = strlen(str);
-    for (int i = 0; i < len; i++) {
-        for (int j = i; j < len; j++) {
-            for (int k = i; k <= j; k++)
+    size_t len = strlen(str);
+    for (size_t i = 0; i < len; i++) {
+        for (size_t j = i; j < len; j++) {
+            for (size_t k = i; k <= j; k++)
                 printf("%c", str[k]);
             printf("\n");
         }
