+++ Q117.c
@@ -4,18 +4,18 @@
 int main() {
     int m, n;
     scanf("%d", &m);
-    int *a = malloc(m * sizeof *a);
+    int *a = malloc((size_t)m * sizeof *a);
     if (!a) return 1;
     for (int i = 0; i < m; i++) {
         scanf("%d", &a[i]);
     }
     scanf("%d", &n);
-    int *b = malloc(n * sizeof *b);
+    int *b = malloc((size_t)n * sizeof *b);
     if (!b) { free(a); return 1; }
     for (int i = 0; i < n; i++) {
         scanf("%d", &b[i]);
     }
-    int *merged = malloc((m + n) * sizeof *merged);
+    int *merged = malloc((size_t)(m + n) * sizeof *merged);
     if (!merged) { free(a); free(b); return 1; }
     int i = 0, j = 0, k = 0;
     while (i < m && j < n) {
