+++ Q17.c
@@ -2,10 +2,10 @@
 #include <math.h>
 
 int main() {
-    float a, b, c, discriminant, root1, root2, realPart, imagPart;
+    double a, b, c, discriminant, root1, root2, realPart, imagPart;
 
     printf("Enter coefficients a, b and c: ");
-    scanf("%f %f %f", &a, &b, &c);
+    scanf("%lf %lf %lf", &a, &b, &c);
 
     discriminant = b * b - 4 * a * c;
